+++ src/main/java/org/mmarini/rl/agents/DLAgentBuilder.java
@@ -1,7 +1,7 @@
 /*
- * Copyright 2026 Marco Marini, marco.marini@mmarini.org
+ * Copyright (c) 2026 Marco Marini, marco.marini@mmarini.org
  *
- * Permission is hereby granted, free of charge, to any person
+ *  Permission is hereby granted, free of charge, to any person
  * obtaining a copy of this software and associated documentation
  * files (the "Software"), to deal in the Software without
  * restriction, including without limitation the rights to use,
@@ -22,7 +22,7 @@
  * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
  * OTHER DEALINGS IN THE SOFTWARE.
  *
- * END OF TERMS AND CONDITIONS
+ *    END OF TERMS AND CONDITIONS
  *
  */
 
@@ -465,9 +465,8 @@ public class DLAgentBuilder {
                         (float) locator.getNode(root).asDouble()
                 ).toMap();
         float beta = (float) Locator.locate(BETA_ID).getNode(root).asDouble();
-        float gamma = (float) Locator.locate(GAMMA_ID).getNode(root).asDouble();
         float avgReward = (float) Locator.locate(AVG_REWARD_ID).getNode(root).asDouble();
-        return DLAgent.create(filePath, network, random, numEpochs, trajectorySize, batchSize, alphas, beta, gamma, avgReward, false);
+        return DLAgent.create(filePath, network, random, numEpochs, trajectorySize, batchSize, alphas, beta, avgReward, false);
     }
 
     private ActivationLayer buildActivationLayer(Locator locator) {